rearrange: create rm module
